refactor: コンポーネントを100行以内に分割
